working task for vegetation
