+++ src/app/core/services/worker_python/tasks/python-scripts/api.py
@@ -9,6 +9,7 @@ import logging
 from stellar_engine.core import geometry, pose_table
 from stellar_engine.core import initialize_study as stellar_initialize_study
 from stellar_engine.core import loads
+from stellar_engine.core.conformity import simulation as conformity_simulation
 from stellar_engine.data import geography
 from stellar_engine.tools import (
     guying,
@@ -16,7 +17,7 @@ from stellar_engine.tools import (
     temperature,
     papoto,
 )
-from stellar_engine.plot import conformity, plot_2d, plot_settings, run_solver, supports_coords
+from stellar_engine.plot import plot_2d, plot_settings, run_solver, supports_coords
 import stellar_engine.plot.obstacles as obst
 
 from mechaphlowers import SectionStudy
@@ -274,7 +275,7 @@ def get_conformity(js_inputs):
     global study
     python_inputs = js_to_python(js_inputs)
     logger.debug(f"python_inputs for get_conformity: {python_inputs}")
-    out =conformity.get_conformity(python_inputs, study)
+    out = conformity_simulation.get_conformity(python_inputs, study)
     logger.debug(f"Conformity output: {out}")
     return out
