+++ stellar-engine/src/stellar_engine/__init__.py
@@ -1,6 +1,6 @@
 import logging
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"
 
 logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())
