+++ stellar-engine/src/stellar_engine/entities/errors.py
