+++ stellar-engine/src/stellar_engine/plot/plot_2d.py
@@ -55,7 +55,9 @@ def get_states_coordinates(
     # Extract formatted obstacles/distances computed before serialization
     obstacles = current_coords.pop("obstacles_formatted", [])
     distances = current_coords.pop("distances_formatted", [])
-    distance_measuring_points = current_coords.pop("distanceMeasuringPoints", [])
+    distance_measuring_points = current_coords.pop(
+        "distanceMeasuringPoints", []
+    )
 
     return {
         "sectionOutput": {"current": current_coords, "base": base_coords},
